+++ src/device/tsp/Tsp/Script/SessionBaseExtensions/NodeMethods_deploy.cs
@@ -1,5 +1,4 @@
 using cc.isr.VI.Pith;
-using System.Diagnostics;
 
 namespace cc.isr.VI.Tsp.Script.SessionBaseExtensions;
 
@@ -18,8 +17,57 @@ public static partial class NodeMethods
     {
         if ( session is null ) throw new ArgumentNullException( nameof( session ) );
         return session.IsNil( nodeNumber, script.VersionGetterElement )
-            ?  Syntax.Tsp.Lua.NilValue
+            ? Syntax.Tsp.Lua.NilValue
             : session.QueryPrintTrimEnd( nodeNumber, script.VersionGetter ) ?? Syntax.Tsp.Lua.NilValue;
+    }
+
+    /// <summary>
+    /// A <see cref="Pith.SessionBase"/> extension method that queries firmware version.
+    /// </summary>
+    /// <remarks>   2025-04-26. </remarks>
+    /// <exception cref="ArgumentNullException">        Thrown when one or more required arguments
+    ///                                                 are null. </exception>
+    /// <exception cref="InvalidOperationException">    Thrown when the requested operation is
+    ///                                                 invalid. </exception>
+    /// <param name="session">      The session. </param>
+    /// <param name="nodeNumber">   The node number. </param>
+    /// <param name="scripts">      The scripts. </param>
+    public static void QueryFirmwareVersion( this SessionBase session, int nodeNumber, ScriptInfoCollection scripts )
+    {
+        if ( session is null ) throw new ArgumentNullException( nameof( session ) );
+        if ( !session.IsSessionOpen ) throw new InvalidOperationException( $"{nameof( session )} is not open." );
+        if ( scripts is null ) throw new ArgumentNullException( nameof( scripts ) );
+
+        foreach ( ScriptInfo script in scripts )
+            script.ActualVersion = session.QueryFirmwareVersion( nodeNumber, script );
+    }
+
+    /// <summary>
+    /// A <see cref="Pith.SessionBase"/> extension method that queries firmware version.
+    /// </summary>
+    /// <remarks>   2025-04-26. </remarks>
+    /// <exception cref="ArgumentNullException">        Thrown when one or more required arguments
+    ///                                                 are null. </exception>
+    /// <exception cref="InvalidOperationException">    Thrown when the requested operation is
+    ///                                                 invalid. </exception>
+    /// <param name="session">      The session. </param>
+    /// <param name="nodeScripts">  The node scripts. </param>
+    public static void QueryFirmwareVersion( this SessionBase session, NodesScriptsCollection nodeScripts )
+    {
+        if ( session is null ) throw new ArgumentNullException( nameof( session ) );
+        if ( !session.IsSessionOpen ) throw new InvalidOperationException( $"{nameof( session )} is not open." );
+        if ( nodeScripts is null ) throw new ArgumentNullException( nameof( nodeScripts ) );
+
+        foreach ( ScriptInfoCollection scriptInfoCollection in nodeScripts.Values )
+        {
+            if ( scriptInfoCollection is not null )
+            {
+                if ( (scriptInfoCollection.NodeNumber == 0) || (scriptInfoCollection.NodeNumber == session.QueryControllerNodeNumber()) )
+                    session.QueryFirmwareVersion( scriptInfoCollection );
+                else
+                    session.QueryFirmwareVersion( scriptInfoCollection.NodeNumber, scriptInfoCollection );
+            }
+        }
     }
 
     /// <summary>   A <see cref="Pith.SessionBase"/> extension method that reads script state. </summary>
@@ -52,7 +100,7 @@ public static partial class NodeMethods
             {
                 embeddedScript.ScriptStatus |= ScriptStatuses.Activated;
                 embeddedScript.ActualVersion = session.QueryFirmwareVersion( nodeNumber, script );
-                if ( session.IsBinaryScript( nodeNumber, script.Title ) )
+                if ( session.IsBinaryScript( nodeNumber, script ) )
                     embeddedScript.ScriptStatus |= ScriptStatuses.BinaryByteCode;
             }